Удалена библиотека json. Параметры бота теперь получаются из переменных окружения. Добавлен Docker
