+++ src/main/java/TelegramBotApp.java
@@ -1,21 +1,21 @@
-import LibBaseDto.DtoBaseBot.Bot;
-import Utils.Parser;
-
-import org.json.JSONObject;
+import LibBaseDto.DtoBaseBot.BotSetting;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
 
 
-public class App {
+public class TelegramBotApp {
     public static void main(String[] args) throws Exception {
 
-        JSONObject setting = Parser.parseJsonFile("BotSetting.json");
-        Bot bot = Bot.setBotInJson(setting);
+        BotSetting botSetting = new BotSetting();
+        botSetting.setName(System.getenv("name"));
+        botSetting.setCreatorId(Integer.parseInt(System.getenv("creatorId")));
+        botSetting.setToken(System.getenv("token"));
+        botSetting.setDbPath(System.getenv("dbPath"));
 
         try {
             TelegramBotsApi botsApi = new TelegramBotsApi(DefaultBotSession.class);
-            botsApi.registerBot(new Telegram(bot));
+            botsApi.registerBot(new Telegram(botSetting));
         } catch (TelegramApiException e) {
             e.printStackTrace();
         }
